Move ssl/cors config and HTTP utils into kinotic-api-gateway
Relocate the browser-facing HTTP configuration and helpers to the module
that owns the web tier. CorsProperties (from core) and SslProperties (from
domain) move into kinotic-api-gateway's api/config; the cors field moves
off KinoticProperties and the ssl field off DomainProperties onto
ApiGatewayProperties. CorsUtil and SslHelper are combined into a single
gateway ApiGatewayUtil.

Wire paths change: kinotic.cors.* -> kinotic.apiGateway.cors.* and
kinotic.domain.ssl.* -> kinotic.apiGateway.ssl.*.

- ApiGatewayProperties gains cors + ssl; gateway consumers read them via
  properties.getApiGateway().get{Cors,Ssl}(). ApiGatewayVertcleFactory no
  longer injects KinoticDomainProperties, and ApiGatewayEndpointInitializer
  drops to the base KinoticProperties, removing gateway's dependence on
  domain properties (the always-on KinoticDomainProperties comment is
  updated to match).
- kinotic-persistence now depends on kinotic-api-gateway and references the
  moved classes / ApiGatewayUtil; OpenApiVertxRouterFactory reads cors from
  KinoticApiGatewayProperties.
- application-development/kubernetes nest cors under apiGateway; helm
  config-map SSL env vars become KINOTIC_APIGATEWAY_SSL_*; values.yaml
  comments updated.
