+++ kinotic-api-gateway/src/main/java/org/kinotic/gateway/api/config/ApiGatewayProperties.java
@@ -34,6 +34,16 @@ public class ApiGatewayProperties {
      */
     private WebServerProperties webServer = new WebServerProperties();
 
+    /**
+     * CORS configuration applied to all Vert.x HTTP servers that expose browser-facing routes.
+     */
+    private CorsProperties cors = new CorsProperties();
+
+    /**
+     * SSL/TLS configuration for all Vert.x HTTP servers.
+     */
+    private SslProperties ssl = new SslProperties();
+
     public ApiGatewayProperties(KinoticProperties kinoticProperties) {
         stomp = new StompServerOptions()
                 .setPort(DEFAULT_STOMP_PORT)
