+++ kinotic-api-gateway/src/main/java/org/kinotic/gateway/api/utils/ApiGatewayUtil.java
@@ -1,12 +1,20 @@
-package org.kinotic.domain.api.utils;
+package org.kinotic.gateway.api.utils;
 
 import io.vertx.core.http.HttpMethod;
+import io.vertx.core.http.HttpServerOptions;
+import io.vertx.core.net.PemKeyCertOptions;
 import io.vertx.ext.web.handler.CorsHandler;
-import org.kinotic.core.api.config.CorsProperties;
+import org.kinotic.gateway.api.config.CorsProperties;
+import org.kinotic.gateway.api.config.SslProperties;
 
 import java.util.Set;
 
-public final class CorsUtil {
+/**
+ * Helpers for configuring the Vert.x HTTP servers fronted by the api-gateway: CORS handling
+ * and PEM-based TLS. Applied to every browser-facing route (STOMP, static web, and the
+ * openapi/graphql routes mounted by other modules).
+ */
+public final class ApiGatewayUtil {
 
     /**
      * The HTTP methods kinotic's HTTP routes accept. Pinned here rather than exposed on
@@ -21,7 +29,7 @@ public final class CorsUtil {
                                                                   HttpMethod.DELETE,
                                                                   HttpMethod.OPTIONS);
 
-    private CorsUtil() {}
+    private ApiGatewayUtil() {}
 
     public static CorsHandler createCorsHandler(CorsProperties properties) {
         String pattern = properties.getAllowedOriginPattern();
@@ -36,5 +44,23 @@ public final class CorsUtil {
             corsHandler.allowCredentials(properties.getAllowCredentials());
         }
         return corsHandler;
+    }
+
+    /**
+     * If SSL is enabled in the given properties, configures the server options
+     * with PEM-based TLS. Otherwise leaves the options unchanged.
+     *
+     * @param options the server options to configure
+     * @param ssl     the SSL properties
+     * @return the same options instance for chaining
+     */
+    public static HttpServerOptions applySsl(HttpServerOptions options, SslProperties ssl) {
+        if (ssl != null && ssl.isEnabled()) {
+            options.setSsl(true)
+                   .setKeyCertOptions(new PemKeyCertOptions()
+                           .setCertPath(ssl.getCertPath())
+                           .setKeyPath(ssl.getKeyPath()));
+        }
+        return options;
     }
 }
