+++ kinotic-api-gateway/src/main/java/org/kinotic/gateway/internal/endpoints/ApiGatewayVertcleFactory.java
@@ -14,9 +14,7 @@ import io.vertx.ext.web.handler.SessionHandler;
 import io.vertx.ext.web.healthchecks.HealthCheckHandler;
 import io.vertx.ext.web.sstore.SessionStore;
 import lombok.RequiredArgsConstructor;
-import org.kinotic.domain.api.utils.SslHelper;
-import org.kinotic.domain.api.utils.CorsUtil;
-import org.kinotic.domain.api.config.KinoticDomainProperties;
+import org.kinotic.gateway.api.utils.ApiGatewayUtil;
 import org.kinotic.domain.api.rest.SuppliesGatewayRoutes;
 import org.kinotic.gateway.api.config.KinoticApiGatewayProperties;
 import org.springframework.stereotype.Component;
@@ -32,7 +30,6 @@ import java.util.List;
 public class ApiGatewayVertcleFactory {
 
     private final KinoticApiGatewayProperties properties;
-    private final KinoticDomainProperties domainProperties;
     private final StompServerHandlerFactory stompServerHandlerFactory;
     private final List<SuppliesGatewayRoutes> gatewayRoutes;
     private final HealthChecks healthChecks;
@@ -44,9 +41,9 @@ public class ApiGatewayVertcleFactory {
 
         // CORS first — the SPA is a different origin from this gateway (portal.kinotic.ai
         // vs api.kinotic.ai in prod, vite's :5173 in dev). They're same-site, so the
-        // SameSite=Lax session cookie flows; credentialed CORS (kinotic.cors.*) lets the
+        // SameSite=Lax session cookie flows; credentialed CORS (kinotic.apiGateway.cors.*) lets the
         // cross-origin login fetch store it. Shared with the openapi/graphql routes.
-        router.route().handler(CorsUtil.createCorsHandler(properties.getCors()));
+        router.route().handler(ApiGatewayUtil.createCorsHandler(properties.getApiGateway().getCors()));
 
         // Health check on the api-gateway port so probes work even when the static
         // web-server (9090) is disabled in KinD/Azure.
@@ -81,12 +78,12 @@ public class ApiGatewayVertcleFactory {
         HttpServerOptions serverOptions = new HttpServerOptions();
         serverOptions.setWebSocketSubProtocols(List.of("v12.stomp"));
         serverOptions.setMaxWebSocketFrameSize(properties.getMaxEventPayloadSize());
-        SslHelper.applySsl(serverOptions, domainProperties.getDomain().getSsl());
+        ApiGatewayUtil.applySsl(serverOptions, properties.getApiGateway().getSsl());
 
         return StompServerVerticleFactory.create(serverOptions, stompServerOptions, stompServerHandlerFactory, router);
     }
 
     public WebServerVerticle createWebServerVerticle(){
-        return new WebServerVerticle(properties.getApiGateway().getWebServer(), domainProperties.getDomain().getSsl());
+        return new WebServerVerticle(properties.getApiGateway().getWebServer(), properties.getApiGateway().getSsl());
     }
 }
