+++ kinotic-domain/src/main/java/org/kinotic/domain_autoconfig/KinoticDomainAutoConfiguration.java
@@ -7,8 +7,8 @@ import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Import;
 
 // KinoticDomainProperties is enabled here (always-on) rather than via KinoticDomainLibrary's
-// @ComponentScan, which is gated by kinotic.disableDomain. The api-gateway reads domain.ssl
-// even when domain logic is disabled, so the properties must remain available.
+// @ComponentScan, which is gated by kinotic.disableDomain, so kinotic.domain.* binds even
+// when domain logic is disabled.
 @AutoConfiguration
 @Import(KinoticDomainLibrary.class)
 @EnableConfigurationProperties(KinoticDomainProperties.class)
