+++ kinotic-persistence/src/main/java/org/kinotic/persistence/internal/endpoints/openapi/OpenApiVerticle.java
@@ -6,8 +6,8 @@ import io.vertx.core.http.HttpServer;
 import io.vertx.core.http.HttpServerOptions;
 import io.vertx.ext.web.Router;
 import lombok.RequiredArgsConstructor;
-import org.kinotic.domain.api.utils.SslHelper;
-import org.kinotic.domain.api.config.SslProperties;
+import org.kinotic.gateway.api.utils.ApiGatewayUtil;
+import org.kinotic.gateway.api.config.SslProperties;
 import org.kinotic.persistence.api.config.PersistenceProperties;
 import org.kinotic.persistence.api.model.EntityDefinition;
 
@@ -28,7 +28,7 @@ public class OpenApiVerticle extends VerticleBase {
     public Future<?> start() throws Exception {
         HttpServerOptions options = new HttpServerOptions();
         options.setMaxHeaderSize(properties.getMaxHttpHeaderSize());
-        SslHelper.applySsl(options, sslProperties);
+        ApiGatewayUtil.applySsl(options, sslProperties);
         server = vertx.createHttpServer(options);
 
         // Begin listening for requests
